Add "site-cookbooks", fix "append" calls, and always deploy chef-solo.
If a "site-cookbooks" directory is found, its contents will selectively
override those in "cookbooks".

Fixed "append" operations, they were failing because they didn't use sudo
and had the arguments for filename and content backwards.

A new "deploy_solo" task ensures that the node has a fresh "solo.rb" and
directory structure. Previously this was created only during "deploy_chef",
which was a problem because "solo.rb" would be out of date and there would
be no tidy way to regenerate it. Now this task is run as part of the
"configure" and "deploy_chef" tasks to ensure there's always a good setup.