add _configure_fabric_for_platform method
Pass in the `platform` property from Ohai to this method, and it will configure fabric for that platform.

Right now it only changes env.shell to `/bin/sh -c` if the platform is freebsd.

We call _configure_fabric_for_platform from the _node_runner method, and from the deploy_chef task